Switch to InnoDB for better data integrity, transactions, and FK
Modifications to the sql import files and minor changes to the main
library in order to properly support the use of foreign keys in Tank Auth.
2 default roles were inserted in the `roles` table in order to satisfy the
FK requirements in the `user_roles` table.

Tip: Feel free to delete the 'user' role. That's just there since you'll
be needing a user role at some point. But keep the 'admin' role since that
is required by the Foreign Key. Rename it if you wish, only the id is
important.